+++ internal/sdk/cloudian/sdk.go
@@ -87,23 +87,23 @@ func fromInternal(g groupInternal) Group {
 	}
 }
 
-type User struct {
-	UserID  string `json:"userId"`
+type UserType string
+
+const (
+	UserTypeSystemAdmin UserType = "SystemAdmin"
+	UserTypeGroupAdmin  UserType = "GroupAdmin"
+	UserTypeStandard    UserType = "User"
+)
+
+type GroupUserID struct {
 	GroupID string `json:"groupId"`
+	UserID  string `json:"userId"`
 }
 
-type userInternal struct {
-	UserID   string `json:"userId"`
-	GroupID  string `json:"groupId"`
-	UserType string `json:"userType"`
-}
-
-func toInternalUser(u User) userInternal {
-	return userInternal{
-		UserID:   u.UserID,
-		GroupID:  u.GroupID,
-		UserType: "User",
-	}
+type User struct {
+	GroupUserID `json:",inline"`
+	UserType    UserType `json:"userType"`
+	CanonicalID string   `json:"canonicalUserId,omitempty"`
 }
 
 // SecurityInfo is the Cloudian API's term for secure credentials
@@ -169,11 +169,11 @@ func (client Client) ListUsers(ctx context.Context, groupID string, userID *stri
 }
 
 // Delete a single user. Errors if the user does not exist.
-func (client Client) DeleteUser(ctx context.Context, user User) error {
+func (client Client) DeleteUser(ctx context.Context, guid GroupUserID) error {
 	resp, err := client.newRequest(ctx).
 		SetQueryParams(map[string]string{
-			"groupId": user.GroupID,
-			"userId":  user.UserID,
+			"groupId": guid.GroupID,
+			"userId":  guid.UserID,
 		}).
 		Delete("/user")
 	if err != nil {
@@ -192,7 +192,7 @@ func (client Client) DeleteUser(ctx context.Context, user User) error {
 // Create a single user of type `User` into a groupId
 func (client Client) CreateUser(ctx context.Context, user User) error {
 	resp, err := client.newRequest(ctx).
-		SetBody(toInternalUser(user)).
+		SetBody(user).
 		Put("/user")
 	if err != nil {
 		return err
@@ -208,13 +208,15 @@ func (client Client) CreateUser(ctx context.Context, user User) error {
 
 // GetUser gets a user. Returns an error even in the case of a user not found.
 // This error can then be checked against ErrNotFound: errors.Is(err, ErrNotFound)
-func (client Client) GetUser(ctx context.Context, user User) (*User, error) {
-	// FIXME: Introduce UserId struct and enrich User
+func (client Client) GetUser(ctx context.Context, guid GroupUserID) (*User, error) {
+	var user User
+
 	resp, err := client.newRequest(ctx).
 		SetQueryParams(map[string]string{
-			"groupId": user.GroupID,
-			"userId":  user.UserID,
+			"groupId": guid.GroupID,
+			"userId":  guid.UserID,
 		}).
+		SetResult(&user).
 		Get("/user")
 	if err != nil {
 		return nil, err
@@ -232,12 +234,12 @@ func (client Client) GetUser(ctx context.Context, user User) (*User, error) {
 }
 
 // CreateUserCredentials creates a new set of credentials for a user.
-func (client Client) CreateUserCredentials(ctx context.Context, user User) (*SecurityInfo, error) {
+func (client Client) CreateUserCredentials(ctx context.Context, guid GroupUserID) (*SecurityInfo, error) {
 	var securityInfo SecurityInfo
 
 	resp, err := client.newRequest(ctx).
 		SetResult(&securityInfo).
-		SetQueryParams(map[string]string{"groupId": user.GroupID, "userId": user.UserID}).
+		SetQueryParams(map[string]string{"groupId": guid.GroupID, "userId": guid.UserID}).
 		Put("/user/credentials")
 	if err != nil {
 		return nil, err
@@ -275,11 +277,11 @@ func (client Client) GetUserCredentials(ctx context.Context, accessKey string) (
 }
 
 // ListUserCredentials fetches all the credentials of a user.
-func (client Client) ListUserCredentials(ctx context.Context, user User) ([]SecurityInfo, error) {
+func (client Client) ListUserCredentials(ctx context.Context, guid GroupUserID) ([]SecurityInfo, error) {
 	var securityInfo []SecurityInfo
 
 	resp, err := client.newRequest(ctx).
-		SetQueryParams(map[string]string{"groupId": user.GroupID, "userId": user.UserID}).
+		SetQueryParams(map[string]string{"groupId": guid.GroupID, "userId": guid.UserID}).
 		SetResult(&securityInfo).
 		Get("/user/credentials/list")
 	if err != nil {
@@ -291,7 +293,7 @@ func (client Client) ListUserCredentials(ctx context.Context, user User) ([]Secu
 		return securityInfo, nil
 	case 204:
 		// Cloudian-API returns 204 if no security credentials found
-		return nil, ErrNotFound
+		return nil, nil
 	default:
 		return nil, fmt.Errorf("error: list credentials unexpected status code: %d", resp.StatusCode())
 	}
@@ -322,7 +324,7 @@ func (client Client) DeleteGroupRecursive(ctx context.Context, groupID string) e
 	}
 
 	for _, user := range users {
-		if err := client.DeleteUser(ctx, user); err != nil {
+		if err := client.DeleteUser(ctx, user.GroupUserID); err != nil {
 			return fmt.Errorf("error deleting user: %w", err)
 		}
 	}
@@ -377,7 +379,7 @@ func (client Client) UpdateGroup(ctx context.Context, group Group) error {
 	case 200:
 		return err
 	default:
-		return fmt.Errorf("Update group unexpected status: %d", resp.StatusCode())
+		return fmt.Errorf("update group unexpected status: %d", resp.StatusCode())
 	}
 }